refactor(acp_native): tighten API — drop test shims, dead variant, fn pointer (#117)
Address self-review code smells flagged on PR #117:

#1 — Three `*_for_test` shims leaked module internals just to bridge the
sibling-module boundary between `acp_native::tests` and the files under
test. Replaced with two visibility tightenings and one test relocation:

- `AcpNativeHandle::alloc_id` is now `pub(super)`. Deleted
  `alloc_id_for_test`. Tests call `alloc_id()` directly.
- `reader::handle_response` is now `pub(super)`. Deleted
  `handle_response_for_test`. Tests call `handle_response(...)` directly.
- The three `close()` multi-session tests moved into `handle.rs::tests`
  as an inline `#[cfg(test)] mod tests` block. Inside the same module
  they construct `AcpNativeHandle` with private field access — no
  `set_session_for_test` setter shim required. Deleted that shim too.

To support tests in multiple files, factored shared fixtures
(`TEST_CFG`, `TEST_REGISTRY`, `test_spec`, `make_core`,
`open_test_session`, `fresh_shared`) into a new
`acp_native/test_fixtures.rs` gated on `#[cfg(test)]`. Both
`acp_native::tests` and `acp_native::handle::tests` import from it.

#2 — `InitPromptStrategy::Deferred` was annotated `#[allow(dead_code)]`
"for future runtimes." YAGNI. Deleted the variant. The enum stays as a
single-variant enum (rather than collapsing to "always immediate"
behavior) so a future driver that genuinely needs to defer can extend
without a wire-shape breaking change. Doc-comment on the enum explains
why.

#4 — `AcpDriverConfig::registry` was `fn() -> &'static AgentRegistry<...>`
wrapping a function-local static. Hoisted each driver's static to module
level (`KIMI_REGISTRY`, `GEMINI_REGISTRY`, `OPENCODE_REGISTRY`,
`TEST_REGISTRY`) and changed the field type to
`&'static AgentRegistry<AcpNativeCore>`. Removed the `(cfg.registry)()`
call indirection at every use site. `AgentRegistry::new` is `const fn`
so this just works.

Verified: cargo fmt --check (clean), cargo test --lib (324 passed),
cargo clippy --lib --tests -- -D warnings (clean).
